build(tests): move out of dist
