+++ __tests__/serializers/combined/fixtures/elementTags.ts
